[Codegen] Do not emit `br` instructions if the insertion point is null
When a block contains a `return` statement, we should not add another
`br` instruction at its end, as it's already terminated. This was
previously achieved by setting the insertion point to null when
encountering a `ResolvedReturnStmt`.

Unfortunately, this "orphan" `br` is still considered a predecessor of
the target basic block (which is shown by `preds = <null operand!>` in
the IR). If using a debug build of LLVM, this triggers a "Uses remain
when a value is destroyed!" assertion in `~BasicBlock`, as the `br` is
never destroyed.